node float coordinates

+++ src/main/java/org/openstreetmap/model/Node.java
@@ -5,12 +5,12 @@ import java.util.GregorianCalendar;
 import java.util.Map;
 
 public final class Node extends OsmBasicType {
-    private final double lat;
-    private final double lon;
+    private final float lat;
+    private final float lon;
 
     private final String toString;
 
-    public Node(BigInteger id, String user, BigInteger uid, GregorianCalendar timestamp, Integer changeset, Integer version, Boolean visible, Map<String, String> tags, double lat, double lon) {
+    public Node(BigInteger id, String user, BigInteger uid, GregorianCalendar timestamp, Integer changeset, Integer version, Boolean visible, Map<String, String> tags, float lat, float lon) {
         super(id, user, uid, timestamp, changeset, version, visible, tags);
         this.lat = lat;
         this.lon = lon;
@@ -18,11 +18,11 @@ public final class Node extends OsmBasicType {
         this.toString = getClass().getSimpleName() + "(id: " + String.valueOf(getId()) + ", lat: " + String.valueOf(getLat()) + ", lon: " + String.valueOf(getLon()) + ")";
     }
 
-    public double getLat() {
+    public float getLat() {
         return lat;
     }
 
-    public double getLon() {
+    public float getLon() {
         return lon;
     }
